# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BidWrenx API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /users/me/photo:
    patch:
      operationId: updateMyPhoto
      tags: [auth]
      summary: Save profile photo object path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhotoInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users/me/verification/request:
    post:
      operationId: requestVerification
      tags: [auth]
      summary: Mechanic submits a verification request
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users/me/verification/documents:
    patch:
      operationId: updateVerificationDocuments
      tags: [auth]
      summary: Save insurance or certification document paths
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentsInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs (browsable by mechanics)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, completed, cancelled]
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new repair job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobStatusUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /bids/{id}/accept:
    patch:
      operationId: acceptBid
      tags: [bids]
      summary: Accept a bid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bid accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bid"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, mechanic]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        warningCount:
          type: integer
        flaggedForReview:
          type: boolean
        suspended:
          type: boolean
        verified:
          type: boolean
        isAdmin:
          type: boolean
        photoUrl:
          type: ["string", "null"]
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        identityVerificationStatus:
          type: string
          enum: [none, pending, approved, rejected]
        insuranceDocumentUploaded:
          type: boolean
        certificationDocumentUploaded:
          type: boolean
        adminVerifiedMechanic:
          type: boolean
        verificationNotes:
          type: ["string", "null"]
        verificationRequestedAt:
          type: ["string", "null"]
        termsAcceptedAt:
          type: ["string", "null"]
        hasPinSet:
          type: boolean
        pinLocked:
          type: boolean
        banned:
          type: boolean
        referralCode:
          type: ["string", "null"]
        specialties:
          type: ["string", "null"]
        legalName:
          type: ["string", "null"]
        displayName:
          type: ["string", "null"]
        username:
          type: ["string", "null"]
        showLegalNamePublicly:
          type: boolean
        mapCity:
          type: ["string", "null"]
        mapState:
          type: ["string", "null"]
        mapLat:
          type: ["number", "null"]
        mapLng:
          type: ["number", "null"]
        serviceRadiusMiles:
          type: integer
        isMobileMechanic:
          type: boolean
        shopType:
          type: ["string", "null"]
        willingToTravelMiles:
          type: ["integer", "null"]
        emergencyAvailable:
          type: boolean
        mapVisible:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, reviewCount, warningCount, flaggedForReview, suspended, verified, isAdmin, emailVerified, phoneVerified, identityVerificationStatus, insuranceDocumentUploaded, certificationDocumentUploaded, adminVerifiedMechanic, hasPinSet, pinLocked, banned, showLegalNamePublicly, serviceRadiusMiles, isMobileMechanic, emergencyAvailable, mapVisible, createdAt]
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        location:
          type: string
        photos:
          type: array
          items:
            type: string
        bidCount:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, description, category, vehicleMake, vehicleModel, vehicleYear, status, location, photos, bidCount, customerId, customerName, createdAt, updatedAt]
    JobDetail:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        location:
          type: string
        photos:
          type: array
          items:
            type: string
        customerId:
          type: integer
        customerName:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        bids:
          type: array
          items:
            $ref: "#/components/schemas/BidDetail"
      required: [id, title, description, category, vehicleMake, vehicleModel, vehicleYear, status, location, photos, customerId, customerName, createdAt, updatedAt, bids]
    JobInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        budgetMin:
          type: number
        budgetMax:
          type: number
        location:
          type: string
        photos:
          type: array
          items:
            type: string
      required: [title, description, category, vehicleMake, vehicleModel, vehicleYear, location]
    JobStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
      required: [status]
...
    Bid:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        mechanicId:
          type: integer
        mechanicName:
          type: string
        amount:
          type: number
        estimatedDays:
          type: integer
        note:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn]
        createdAt:
          type: string
      required: [id, jobId, mechanicId, mechanicName, amount, estimatedDays, status, createdAt]
    BidDetail:
      type: object
      properties:
        id:
          type: integer
        mechanicId:
          type: integer
        mechanicName:
          type: string
        mechanicRating:
          type: ["number", "null"]
        mechanicReviewCount:
          type: integer
        mechanicPhotoUrl:
          type: ["string", "null"]
        mechanicAdminVerified:
          type: boolean
        amount:
          type: number
        estimatedDays:
          type: integer
        note:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, accepted, rejected, withdrawn]
        createdAt:
          type: string
      required: [id, mechanicId, mechanicName, mechanicReviewCount, mechanicAdminVerified, amount, estimatedDays, status, createdAt]
...
    UpdatePhotoInput:
      type: object
      properties:
        photoUrl:
          type: string
      required: [photoUrl]
    UpdateDocumentsInput:
      type: object
      properties:
        insuranceDocumentPath:
          type: ["string", "null"]
        certificationDocumentPath:
          type: ["string", "null"]
...