# lib/api-zod/src/generated/types/updateDocumentsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BidWrenx API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateDocumentsInput {
  /** @nullable */
  insuranceDocumentPath?: string | null;
  /** @nullable */
  certificationDocumentPath?: string | null;
}